+++ sdk/keyvault/keyvault-certificates/test/internal/challengeAuthenticationCallbacks.spec.ts
@@ -17,7 +17,7 @@ describe("Challenge based authentication tests", function () {
   let challengeCallbacks: ChallengeCallbacks;
 
   beforeEach(() => {
-    request = createPipelineRequest({ url: "https://foo.bar" });
+    request = createPipelineRequest({ url: "https://myvault.vault.azure.net" });
     challengeCallbacks = createChallengeCallbacks();
   });
 
@@ -59,7 +59,7 @@ describe("Challenge based authentication tests", function () {
         request,
         response: {
           headers: createHttpHeaders({
-            "WWW-Authenticate": `Bearer scope="cae_scope"`,
+            "WWW-Authenticate": `Bearer resource="https://vault.azure.net"`,
           }),
           request,
           status: 200,
@@ -118,7 +118,7 @@ describe("Challenge based authentication tests", function () {
         request,
         response: {
           headers: createHttpHeaders({
-            "WWW-Authenticate": `Bearer resource="cae_scope"`,
+            "WWW-Authenticate": `Bearer resource="https://vault.azure.net"`,
           }),
           request,
           status: 200,
@@ -126,7 +126,94 @@ describe("Challenge based authentication tests", function () {
         scopes: [],
       });
 
-      assert.sameMembers(getAccessTokenScopes, ["cae_scope/.default"]);
+      assert.sameMembers(getAccessTokenScopes, ["https://vault.azure.net/.default"]);
+    });
+
+    it("throws if the resource is not a valid URL", async () => {
+      await assert.isRejected(
+        challengeCallbacks.authorizeRequestOnChallenge!({
+          getAccessToken: () => Promise.resolve(null),
+          request,
+          response: {
+            headers: createHttpHeaders({
+              "WWW-Authenticate": `Bearer resource="invalid_scope"`,
+            }),
+            request,
+            status: 200,
+          },
+          scopes: [],
+        }),
+        `The challenge contains invalid scope 'invalid_scope/.default'`
+      );
+    });
+
+    it("throws if the resource URI host does not match the request by default", async () => {
+      await assert.isRejected(
+        challengeCallbacks.authorizeRequestOnChallenge!({
+          getAccessToken: () => Promise.resolve(null),
+          request: createPipelineRequest({ url: "https://foo.bar" }),
+          response: {
+            headers: createHttpHeaders({
+              "WWW-Authenticate": `Bearer resource="https://vault.azure.net"`,
+            }),
+            request,
+            status: 200,
+          },
+          scopes: [],
+        }),
+        "Challenge resource host 'vault.azure.net' does not match request domain"
+      );
+    });
+
+    it("throws if the request host is a prefix, but not a subdomain, of the resource URI host", async () => {
+      await assert.isRejected(
+        challengeCallbacks.authorizeRequestOnChallenge!({
+          getAccessToken: () => Promise.resolve(null),
+          request: createPipelineRequest({ url: "https://myvault.azure.net" }),
+          response: {
+            headers: createHttpHeaders({
+              "WWW-Authenticate": `Bearer resource="https://vault.azure.net"`,
+            }),
+            request,
+            status: 200,
+          },
+          scopes: [],
+        }),
+        "Challenge resource host 'vault.azure.net' does not match request domain"
+      );
+    });
+
+    it("does not throw if the resource URI matches the request", async () => {
+      await challengeCallbacks.authorizeRequestOnChallenge!({
+        getAccessToken: () => Promise.resolve(null),
+        request: createPipelineRequest({ url: "https://myvault.vault.azure.net" }),
+        response: {
+          headers: createHttpHeaders({
+            "WWW-Authenticate": `Bearer resource="https://vault.azure.net"`,
+          }),
+          request,
+          status: 200,
+        },
+        scopes: [],
+      });
+    });
+
+    it("does not throw if the resource URI host does not match the request but verifyChallengeResource is false", async () => {
+      challengeCallbacks = createChallengeCallbacks({
+        disableChallengeResourceVerification: true,
+      });
+      await challengeCallbacks.authorizeRequestOnChallenge!({
+        getAccessToken: () => Promise.resolve(null),
+        request: createPipelineRequest({ url: "https://foo.bar" }),
+        response: {
+          headers: createHttpHeaders({
+            "WWW-Authenticate": `Bearer resource="https://vault.azure.net"`,
+          }),
+          request,
+          status: 200,
+        },
+        scopes: [],
+      });
     });
 
     it("passes the tenantId if provided", async () => {
@@ -142,7 +229,7 @@ describe("Challenge based authentication tests", function () {
         request,
         response: {
           headers: createHttpHeaders({
-            "WWW-Authenticate": `Bearer scope="cae_scope" authorization="http://login.windows.net/${expectedTenantId}"`,
+            "WWW-Authenticate": `Bearer resource="https://vault.azure.net" authorization="http://login.windows.net/${expectedTenantId}"`,
           }),
           request,
           status: 200,
@@ -161,7 +248,7 @@ describe("Challenge based authentication tests", function () {
         request,
         response: {
           headers: createHttpHeaders({
-            "WWW-Authenticate": `Bearer scope="cae_scope"`,
+            "WWW-Authenticate": `Bearer resource="https://vault.azure.net"`,
           }),
           request,
           status: 200,
@@ -179,7 +266,7 @@ describe("Challenge based authentication tests", function () {
         request,
         response: {
           headers: createHttpHeaders({
-            "WWW-Authenticate": `Bearer scope="cae_scope"`,
+            "WWW-Authenticate": `Bearer resource="https://vault.azure.net"`,
           }),
           request,
           status: 200,