+++ sdk/keyvault/keyvault-common/src/challengeBasedAuthenticationPolicy.ts
@@ -37,6 +37,32 @@ type ChallengeState =
       scopes: string[];
     };
 
+export interface CreateChallengeCallbacksOptions {
+  /**
+   * Whether to disable verification that the challenge resource matches the Key Vault or Managed HSM domain.
+   *
+   * Defaults to false.
+   */
+  disableChallengeResourceVerification?: boolean;
+}
+
+function verifyChallengeResource(scope: string, request: PipelineRequest): void {
+  let scopeAsUrl: URL;
+  try {
+    scopeAsUrl = new URL(scope);
+  } catch (e) {
+    throw new Error(`The challenge contains invalid scope '${scope}'`);
+  }
+
+  const requestUrl = new URL(request.url);
+
+  if (!requestUrl.hostname.endsWith(`.${scopeAsUrl.hostname}`)) {
+    throw new Error(
+      `Challenge resource host '${scopeAsUrl.hostname}' does not match request domain`
+    );
+  }
+}
+
 /**
  * @internal
  *
@@ -50,8 +76,11 @@ type ChallengeState =
  *
  * Following the first request of a client, follow-up requests will get the cached token
  * if possible.
+ *
  */
-export function createChallengeCallbacks(): ChallengeCallbacks {
+export function createChallengeCallbacks({
+  disableChallengeResourceVerification,
+}: CreateChallengeCallbacksOptions = {}): ChallengeCallbacks {
   let challengeState: ChallengeState = { status: "none" };
 
   function requestToOptions(request: PipelineRequest): GetTokenOptions {
@@ -115,6 +144,10 @@ export function createChallengeCallbacks(): ChallengeCallbacks {
 
     if (!scope) {
       throw new Error("Missing scope.");
+    }
+
+    if (!disableChallengeResourceVerification) {
+      verifyChallengeResource(scope, request);
     }
 
     const accessToken = await options.getAccessToken([scope], {