+++ sdk/keyvault/keyvault-secrets/CHANGELOG.md
@@ -1,5 +1,12 @@
 # Release History
 
+## 4.6.0 (2022-09-20)
+
+### Breaking Changes
+
+- Verify the challenge resource matches the vault domain.
+  This should affect few customers who can set `disableChallengeResourceVerification` in the options bag to `true` to disable.
+
 ## 4.5.1 (2022-08-15)
 
 ### Bugs Fixed